requirements.txt
absl-py==2.0.0
accelerate==0.24.1
altair==5.1.2
astunparse==1.6.3
attrs==23.1.0
blinker==1.7.0
Brotli @ file:///C:/Windows/Temp/abs_63l7912z0e/croots/recipe/brotli-split_1659616056886/work
cachetools==5.3.2
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
cycler==0.12.1
decorator==4.4.2
exceptiongroup==1.1.3
expecttest==0.1.6
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
Flask==3.0.0
flatbuffers==23.5.26
flwr==1.5.0
fonttools==4.44.1
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
huggingface-hub==0.19.3
hypothesis==6.88.4
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.1
iterators==0.0.2
itsdangerous==2.1.2
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keras==2.14.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci/markupsafe_1654508077284/work
matplotlib==3.8.1
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
ml-dtypes==0.2.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
openai-whisper==20231106
opt-einsum==3.3.0
optree==0.10.0
packaging==23.2
pandas==2.1.3
Pillow @ file:///C:/b/abs_20ztcm8lgk/croot/pillow_1696580089746/work
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.6
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
pydeck==0.8.1b0
Pygments==2.16.1
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.1.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
referencing==0.31.0
regex==2023.10.3
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.0
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
streamlit==1.28.2
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tiktoken==0.5.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.2
types-dataclasses==0.6.6
typing_extensions @ file:///C:/b/abs_213vg2cd1l/croot/typing_extensions_1690297804941/work
tzdata==2023.3
tzlocal==5.2
urllib3 @ file:///C:/b/abs_9cmlsrm3ys/croot/urllib3_1698257595508/work
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt==1.14.1
zipp==3.17.0
